Fixed code by comments, separated input component
