+++ src/components/Footer/Form/index.js
@@ -1,5 +1,5 @@
 /* eslint-disable react/jsx-props-no-spreading */
-import * as F from '@components/Field/styled';
+import InputMail from '@components/InputMail/index';
 import React from 'react';
 
 import * as S from './styled';
@@ -12,38 +12,9 @@ export default function Form() {
             id="ContactFooter"
             accept-charset="UTF-8"
         >
-            <S.Inner>
-                <S.Field>
-                    <F.Input
-                        id="newsletterForm--footer"
-                        type="email"
-                        name="contact[email]"
-                        defaultValue=""
-                        aria-required="true"
-                        autoCorrect="off"
-                        autoCapitalize="off"
-                        autoComplete="email"
-                        placeholder="Email"
-                        data-validate="email|required"
-                        data-field-value=""
-                        data-field-submit=""
-                    />
-                    <F.Label htmlFor="newsletterForm--footer">
-                        Enter your email to subscribe
-                    </F.Label>
-                </S.Field>
-                <S.Button
-                    defaultValue="Enter"
-                    type="button"
-                    name="commit"
-                    id="Subscribe"
-                    aria-label="Subscribe"
-                >
-                    <S.Icon />
-                </S.Button>
-            </S.Inner>
+            <InputMail />
             <S.Success>
-                <S.PSuccess>Thank you for your message!</S.PSuccess>
+                <S.SuccessText>Thank you for your message!</S.SuccessText>
             </S.Success>
         </S.Form>
     );
