+++ src/components/Header/styled.js
@@ -10,7 +10,6 @@ export const Header = styled.header`
     position: fixed;
     width: 100%;
     height: var(--headerOffset);
-    background-color: ${({ isScrolled }) => (isScrolled ? '' : '')};
     color: var(--header-color, ${({ theme }) => theme.colors.brand});
     transition: background-color 0.3s ease, color 0.3s ease,
         transform 0.3s linear, -webkit-transform 0.3s linear;
